fix(gym): 헬스장 빌드 오류 수정 /GYMMATE-86
